Formulario-definitivo/requirements.txt
fastapi
uvicorn
requests
jinja2
reportlab
python-multipart
pandas
python-dotenv
